pyproject.toml
[project]
name = "online-poll-system-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=5.2.10",
    "django-allauth>=65.14.0",
    "django-anymail[brevo]>=14.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.0.1",
    "langchain>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.8",
    "langchain-postgres>=0.0.16",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.2",
    "pyjwt[crypto]>=2.11.0",
    "qrcode>=8.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sentry-sdk[django]>=2.52.0",
    "shortuuid>=1.0.13",
    "strawberry-graphql>=0.291.0",
    "strawberry-graphql-django>=0.75.0",
    "types-qrcode>=8.2.0.20250914",
    "uvicorn>=0.40.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.2",
    "factory-boy>=3.3.3",
    "pre-commit>=4.5.1",
    "pytest-django>=4.11.1",
    "pytest-cov>=4.1.0",
    "ruff>=0.14.14",
    "pytest-mock>=3.15.1",
    "mypy>=1.8.0",
    "django-stubs>=4.2.7",
    "djangorestframework-stubs>=3.14.5",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12.12",
    "types-python-dateutil>=2.8.19.14",
    "safety>=3.0.1",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--nomigrations --reuse-db"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["*/migrations/*", "manage.py", "config/settings/*"]

[tool.ruff.lint]
# E: pycodestyle, F: Pyflakes, I: isort, U: pyupgrade, N: pep8-naming
# S: bandit (security), B: bugbear, A: flake8-builtins, C4: comprehensions
# T20: flake8-print, PT: pytest-style, DJ: flake8-django, C90: complexity
select = ["E", "F", "I", "U", "N", "S", "B", "A", "C4", "T20", "PT", "DJ", "C90"]
ignore = [
    "S101", # Allow assert for pytest
    "S104", # Allow 0.0.0.0 for Docker local dev
    "S105", # Ignore false positive hardcoded password for settings strings
    "DJ001", # Avoid enforcing null checks on string fields (contentious)
    "DJ008", # Allow functions in models.__str__
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "environ.*",
    "allauth.*",
    "qrcode.*"
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.test"
